web-extension: Handle password forms submitted via JS
Currently, Epiphany relies on the DOM submit event to detect if a form
has been submitted. However, this fails if the form has been submitted
via JavaScript, as the submit event is not emitted in this case. We
can't detect this using DOM alone, so use the new WebKitWebPage signal
that has been provided for this purpose.

https://bugzilla.gnome.org/show_bug.cgi?id=742573